move utils.js
